+++ server/src-lib/Hasura/Server/Auth.hs
@@ -24,8 +24,11 @@ module Hasura.Server.Auth
   , getUserInfoWithExpTime_
   ) where
 
-import qualified Control.Concurrent.Async.Lifted.Safe as LA
-import           Control.Concurrent.Extended          (forkImmortal)
+import           Hasura.Prelude
+
+import           Control.Concurrent.Extended          (ForkableMonadIO, forkManagedT)
+import           Control.Monad.Trans.Managed              (ManagedT)
+import           Control.Monad.Morph                  (hoist)
 import           Control.Monad.Trans.Control          (MonadBaseControl)
 import           Data.IORef                           (newIORef)
 import           Data.Time.Clock                      (UTCTime)
@@ -38,7 +41,6 @@ import qualified Network.HTTP.Client                  as H
 import qualified Network.HTTP.Types                   as N
 
 import           Hasura.Logging
-import           Hasura.Prelude
 import           Hasura.RQL.Types
 
 import           Hasura.Server.Auth.JWT               hiding (processJwt_)
@@ -101,9 +103,7 @@ data AuthMode
 -- This must only be run once, on launch.
 setupAuthMode
   :: ( HasVersion
-     , MonadIO m
-     , MonadBaseControl IO m
-     , LA.Forall (LA.Pure m)
+     , ForkableMonadIO m
      , Tracing.HasReporter m
      )
   => Maybe AdminSecretHash
@@ -112,7 +112,7 @@ setupAuthMode
   -> Maybe RoleName
   -> H.Manager
   -> Logger Hasura
-  -> ExceptT Text m AuthMode
+  -> ExceptT Text (ManagedT m) AuthMode
 setupAuthMode mAdminSecretHash mWebHook mJwtSecret mUnAuthRole httpManager logger =
   case (mAdminSecretHash, mWebHook, mJwtSecret) of
     (Just hash, Nothing,   Nothing)      -> return $ AMAdminSecret hash mUnAuthRole
@@ -147,41 +147,37 @@ setupAuthMode mAdminSecretHash mWebHook mJwtSecret mUnAuthRole httpManager logge
     -- mkJwtCtx :: HasVersion => JWTConfig -> m JWTCtx
     mkJwtCtx
       :: ( HasVersion
-         , MonadIO m
-         , MonadBaseControl IO m
-         , LA.Forall (LA.Pure m)
+         , ForkableMonadIO m
          , Tracing.HasReporter m
          )
       => JWTConfig
-      -> ExceptT T.Text m JWTCtx
+      -> ExceptT T.Text (ManagedT m) JWTCtx
     mkJwtCtx JWTConfig{..} = do
       jwkRef <- case jcKeyOrUrl of
         Left jwk  -> liftIO $ newIORef (JWKSet [jwk])
         Right url -> getJwkFromUrl url
-      return $ JWTCtx jwkRef jcAudience jcIssuer jcClaims
+      return $ JWTCtx jwkRef jcAudience jcIssuer jcClaims jcAllowedSkew
       where
         -- if we can't find any expiry time for the JWK (either in @Expires@ header or @Cache-Control@
         -- header), do not start a background thread for refreshing the JWK
         getJwkFromUrl url = do
           ref <- liftIO $ newIORef $ JWKSet []
-          maybeExpiry <- withJwkError $ Tracing.runTraceT "jwk init" $ updateJwkRef logger httpManager url ref
+          maybeExpiry <- hoist lift $ withJwkError $ Tracing.runTraceT "jwk init" $ updateJwkRef logger httpManager url ref
           case maybeExpiry of
             Nothing   -> return ref
             Just time -> do
-              void . lift $ forkImmortal "jwkRefreshCtrl" logger $
+              void . lift $ forkManagedT "jwkRefreshCtrl" logger $
                 jwkRefreshCtrl logger httpManager url ref (convertDuration time)
               return ref
 
         withJwkError act = do
           res <- runExceptT act
-          case res of
-            Right r -> return r
-            Left err  -> case err of
-              -- when fetching JWK initially, except expiry parsing error, all errors are critical
-              JFEHttpException _ msg  -> throwError msg
-              JFEHttpError _ _ _ e    -> throwError e
-              JFEJwkParseError _ e    -> throwError e
-              JFEExpiryParseError _ _ -> return Nothing
+          onLeft res $ \case
+            -- when fetching JWK initially, except expiry parsing error, all errors are critical
+            JFEHttpException _ msg -> throwError msg
+            JFEHttpError _ _ _ e -> throwError e
+            JFEJwkParseError _ e -> throwError e
+            JFEExpiryParseError _ _ -> return Nothing
 
 getUserInfo
   :: (HasVersion, MonadIO m, MonadBaseControl IO m, MonadError QErr m, Tracing.MonadTrace m)