+++ server/src-lib/Hasura/Server/SchemaUpdate.hs
@@ -15,6 +15,7 @@ import           Hasura.Server.App           (SchemaCacheRef (..), withSCUpdate)
 import           Hasura.Server.Init          (InstanceId (..))
 import           Hasura.Server.Logging
 
+import           Control.Monad.Trans.Managed (ManagedT)
 import           Data.Aeson
 import           Data.Aeson.Casing
 import           Data.Aeson.TH
@@ -81,7 +82,7 @@ $(deriveToJSON
 
 -- | An IO action that enables metadata syncing
 startSchemaSyncThreads
-  :: (MonadIO m)
+  :: C.ForkableMonadIO m
   => SQLGenCtx
   -> PG.PGPool
   -> Logger Hasura
@@ -89,7 +90,7 @@ startSchemaSyncThreads
   -> SchemaCacheRef
   -> InstanceId
   -> Maybe UTC.UTCTime
-  -> m (Immortal.Thread, Immortal.Thread)
+  -> ManagedT m (Immortal.Thread, Immortal.Thread)
   -- ^ Returns: (listener handle, processor handle)
 startSchemaSyncThreads sqlGenCtx pool logger httpMgr cacheRef instanceId cacheInitTime = do
   -- only the latest event is recorded here
@@ -97,12 +98,12 @@ startSchemaSyncThreads sqlGenCtx pool logger httpMgr cacheRef instanceId cacheIn
   updateEventRef <- liftIO $ STM.newTVarIO Nothing
 
   -- Start listener thread
-  lTId <- liftIO $ C.forkImmortal "SchemeUpdate.listener" logger $
+  lTId <- C.forkManagedT "SchemeUpdate.listener" logger . liftIO $
     listener sqlGenCtx pool logger httpMgr updateEventRef cacheRef instanceId cacheInitTime
   logThreadStarted TTListener lTId
 
   -- Start processor thread
-  pTId <- liftIO $ C.forkImmortal "SchemeUpdate.processor" logger $
+  pTId <- C.forkManagedT "SchemeUpdate.processor" logger . liftIO $
     processor sqlGenCtx pool logger httpMgr updateEventRef cacheRef instanceId
   logThreadStarted TTProcessor pTId